+++ src/main/java/model/repositories/CropRepository.java
@@ -3,6 +3,6 @@ package model.repositories;
 import model.Crop;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CropRepository extends CrudRepository<Crop, Long> {
+public interface CropRepository extends CrudRepository<Crop, Short> {
 
 }
